requirements.txt
asgiref==3.9.1
Django==5.2.5
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
psycopg2-binary>=2.9.11
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
